Added tests with included restrictions

+++ app/controllers/recipes_controller.rb
@@ -1,6 +1,6 @@
 class RecipesController < ApplicationController
   
-  before_action :set_recipe, only: [:show, :update, :edit]
+  before_action :set_recipe, only: [:show, :update, :edit, :destroy]
   
   def index
     @recipes = Recipe.paginate(page: params[:page], per_page: 5)
@@ -26,7 +26,10 @@ class RecipesController < ApplicationController
   end
   
   def edit
-    
+    if @recipe.chef != current_chef
+      flash[:danger] = "You are not allowed to edit recipes that don't belong to you!"
+      redirect_to recipe_path(@recipe)
+    end
   end
   
   def update
@@ -39,6 +42,10 @@ class RecipesController < ApplicationController
   end
   
   def destroy
+    if @recipe.chef != current_chef
+      flash[:danger] = "You are not allowed to delete recipes that don't belong to you!"
+      redirect_to recipe_path(@recipe)
+    end
     set_recipe.destroy
     flash[:success] = "Recipe deleted!"
     redirect_to recipes_path
